requirements.txt
blockchain==1.4.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
coinbase==2.1.0
cryptocompare==0.7.0
cryptography==2.9.2
decorator==4.4.2
emoji==0.5.4
enum-compat==0.0.3
Flask==1.1.2
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
pycparser==2.20
pycryptodome==3.9.7
pyTelegramBotAPI==3.6.7
python-telegram-bot==12.6.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
tornado==6.0.4
urllib3==1.25.9
Werkzeug==1.0.1
python-dotenv==0.13.0